--- flickit-assessment-common/src/main/java/org/flickit/assessment/common/application/domain/notification/SendNotificationPort.java
@@ -1,6 +1,0 @@
-package org.flickit.assessment.common.application.domain.notification;
-
-public interface SendNotificationPort {
-
-    void send(NotificationEnvelope envelope);
-}
