+++ flickit-assessment-core/src/test/java/org/flickit/assessment/core/application/service/assessmentinvite/InviteAssessmentUserServiceTest.java
@@ -1,7 +1,7 @@
 package org.flickit.assessment.core.application.service.assessmentinvite;
 
 import org.flickit.assessment.common.application.domain.assessment.AssessmentAccessChecker;
-import org.flickit.assessment.common.application.port.SendEmailPort;
+import org.flickit.assessment.common.application.port.out.SendEmailPort;
 import org.flickit.assessment.common.config.AppSpecProperties;
 import org.flickit.assessment.common.exception.AccessDeniedException;
 import org.flickit.assessment.common.exception.ResourceNotFoundException;
