+++ flickit-assessment-users/src/main/java/org/flickit/assessment/users/application/service/expertgroupaccess/InviteExpertGroupMemberService.java
@@ -3,7 +3,7 @@ package org.flickit.assessment.users.application.service.expertgroupaccess;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.flickit.assessment.common.application.MessageBundle;
-import org.flickit.assessment.common.application.port.SendEmailPort;
+import org.flickit.assessment.common.application.port.out.SendEmailPort;
 import org.flickit.assessment.common.config.AppSpecProperties;
 import org.flickit.assessment.common.exception.AccessDeniedException;
 import org.flickit.assessment.common.exception.ResourceAlreadyExistsException;
